Renaming to @types
