+++ packages/crypto-server/src/routes/v1/generate.ts
@@ -1,6 +1,6 @@
 import * as fastify from 'fastify';
 import generate from '@sebastienrousseau/crypto-lib/dist/lib/generate';
-import { generateHeaders } from '../../types/types';
+import { generateHeaders } from '../../@types/types';
 
 export default (app: fastify.FastifyInstance) => {
   app.get<{
