+++ packages/crypto-server/src/routes/v1/revoke.ts
@@ -1,6 +1,6 @@
 import * as fastify from 'fastify';
 import revoke from '@sebastienrousseau/crypto-lib/dist/lib/revoke';
-import { encryptionHeaders } from '../../types/types';
+import { encryptionHeaders } from '../../@types/types';
 
 export default (app: fastify.FastifyInstance) => {
   app.get<{
